split up cohortslist
